small bugfixes

+++ Tests/App/AppKernel.php
@@ -53,6 +53,10 @@ class AppKernel extends Kernel
      */
     private function removeDirectory($dir)
     {
+        if (!is_dir($dir)) {
+            return;
+        }
+
         $files = array_diff(scandir($dir), array('.', '..'));
         foreach ($files as $file) {
             (is_dir("$dir/$file")) ? $this->removeDirectory("$dir/$file") : unlink("$dir/$file");
